src/requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
decorator==4.0.11
Django==1.10.6
django-admin-honeypot==1.0.0
django-appconf==1.0.2
django-bootstrap3==8.2.1
django-compressor==2.1.1
django-dynamic-fixture==1.9.3
django-environ==0.4.1
django-extensions==1.7.7
django-robots==3.0
django-solo==1.1.2
django-webtest==1.9.1
django-wysiwyg-redactor==0.5.0
easy-thumbnails==2.3
enum34==1.1.6
funcsigs==1.0.2
ipaddress==1.0.18
ipython==5.3.0
ipython-genutils==0.2.0
mock==2.0.0
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
rcssmin==1.0.6
rjsmin==1.0.12
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.2
WebTest==2.0.27
